requirements.txt
CVXcanon==0.0.23.3
Django==1.8.6
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
beautifulsoup4==4.4.1
coverage==4.0.3
cycler==0.9.0
django-admin-bootstrapped==2.5.7
django-bulk-update==1.1.10
django-extensions==1.5.9
ecos==2.0.4
matplotlib==1.5.1
mpld3==0.2
numpy==1.10.1
pygraphviz==1.3.1
pyparsing==2.0.7
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
scikit-learn==0.17
scipy==0.16.1
scs==1.1.7
six==1.10.0
toolz==0.7.4
wikipedia==1.4.0
